+++ quam/core/quam_classes.py
@@ -73,7 +73,7 @@ def convert_dict_and_list(value, cls_or_obj=None, attr=None):
     """Convert a dict or list to a QuamDict or QuamList if possible."""
     if isinstance(value, dict):
         value_annotation = _get_value_annotation(cls_or_obj=cls_or_obj, attr=attr)
-        return QuamDict(**value, value_annotation=value_annotation)
+        return QuamDict(value, value_annotation=value_annotation)
     elif type(value) == list:
         value_annotation = _get_value_annotation(cls_or_obj=cls_or_obj, attr=attr)
         return QuamList(value, value_annotation=value_annotation)
@@ -497,7 +497,11 @@ class QuamBase(ReferenceClass):
                 self, reference, root=self._root
             )
         except ValueError as e:
-            warnings.warn(str(e))
+            try:
+                ref = f"{self.__class__.__name__}: {self.get_reference()}"
+            except Exception:
+                ref = self.__class__.__name__
+            warnings.warn(f"Could not get reference {reference} from {ref}.\n{str(e)}")
             return reference
 
     def print_summary(self, indent: int = 0):
@@ -743,7 +747,11 @@ class QuamDict(UserDict, QuamBase):
         try:
             return self[key]
         except KeyError as e:
-            raise AttributeError(key) from e
+            try:
+                repr = f"{self.__class__.__name__}: {self.get_reference()}"
+            except Exception:
+                repr = self.__class__.__name__
+            raise AttributeError(f'{repr} has no attribute "{key}"') from e
 
     def __setattr__(self, key, value):
         if key in ["data", "parent", "config_settings", "_initialized"]:
@@ -757,7 +765,13 @@ class QuamDict(UserDict, QuamBase):
             try:
                 elem = self._get_referenced_value(elem)
             except ValueError as e:
-                raise KeyError(str(e)) from e
+                try:
+                    repr = f"{self.__class__.__name__}: {self.get_reference()}"
+                except Exception:
+                    repr = self.__class__.__name__
+                raise KeyError(
+                    f"Could not get referenced value {elem} from {repr}"
+                ) from e
         return elem
 
     # Overriding methods from UserDict
@@ -786,6 +800,28 @@ class QuamDict(UserDict, QuamBase):
     ) -> Dict[str, Any]:
         # TODO implement reference kwargs
         return self.data
+
+    def get_attr_name(self, attr_val: Any) -> Union[str, int]:
+        """Get the name of an attribute that matches the value.
+
+        Args:
+            attr_val: The value of the attribute.
+
+        Returns:
+            The name of the attribute. This can also be an int depending on the dict key
+
+        Raises:
+            AttributeError if not found.
+        """
+        for attr_name in self._get_attr_names():
+            if attr_name in self and self[attr_name] is attr_val:
+                return attr_name
+        else:
+            raise AttributeError(
+                "Could not find name corresponding to attribute.\n"
+                f"attribute: {attr_val}\n"
+                f"obj: {self}"
+            )
 
     def _val_matches_attr_annotation(self, attr: str, val: Any) -> bool:
         """Check whether the type of an attribute matches the annotation.