+++ tests/instantiation/test_instantiation.py
@@ -1,13 +1,10 @@
 import pytest
-from typing import List, Optional
+from typing import List, Literal, Optional
 
 from quam.core import QuamRoot, QuamComponent, quam_dataclass
 from quam.examples.superconducting_qubits.components import Transmon
 from quam.core.quam_instantiation import *
-from quam.utils import (
-    get_dataclass_attr_annotations,
-    validate_obj_type,
-)
+from quam.utils import get_dataclass_attr_annotations
 
 
 def test_get_dataclass_attributes():
@@ -73,58 +70,6 @@ def test_get_dataclass_attributes_subdataclass():
         "d": str,
         "elem": float,
     }
-
-
-def test_validate_standard_types():
-    validate_obj_type(1, int)
-    validate_obj_type(1.0, float)
-    validate_obj_type("hello", str)
-    validate_obj_type(":reference", str)
-    validate_obj_type([1, 2, 3], list)
-    validate_obj_type((1, 2, 3), tuple)
-    validate_obj_type({"a": 1, "b": 2}, dict)
-    validate_obj_type(True, bool)
-    validate_obj_type(None, type(None))
-
-    with pytest.raises(TypeError):
-        validate_obj_type(1, str)
-    with pytest.raises(TypeError):
-        validate_obj_type("hello", int)
-
-
-def test_validate_type_exceptions():
-    validate_obj_type("#/reference", int)
-    validate_obj_type("#/reference", str)
-    validate_obj_type("#./reference", int)
-    validate_obj_type("#./reference", str)
-    validate_obj_type("#../reference", int)
-    validate_obj_type("#../reference", str)
-
-    validate_obj_type(None, int)
-    validate_obj_type(None, str)
-
-
-def test_validate_typing_list():
-    validate_obj_type([1, 2, 3], List[int])
-    with pytest.raises(TypeError):
-        validate_obj_type([1, 2, 3], List[str])
-
-    validate_obj_type([1, 2, 3], List)
-    validate_obj_type(["a", "b", "c"], List)
-    validate_obj_type(["a", "b", "c"], List[str])
-    with pytest.raises(TypeError):
-        validate_obj_type(["a", "b", "c"], List[int])
-
-
-def test_validate_typing_dict():
-    validate_obj_type({"a": 1, "b": 2}, dict)
-    validate_obj_type({"a": 1, "b": 2}, Dict[str, int])
-    with pytest.raises(TypeError):
-        validate_obj_type({"a": 1, "b": 2}, Dict[str, str])
-
-    validate_obj_type("#/reference", Dict[str, int])
-    validate_obj_type("#./reference", Dict[str, int])
-    validate_obj_type("#../reference", Dict[str, int])
 
 
 @quam_dataclass
@@ -347,6 +292,28 @@ def test_instantiate_sublist():
     class TestQuamSubList(QuamComponent):
         sublist: List[List[float]]
 
-    obj = instantiate_quam_class(TestQuamSubList, {"sublist": [[1,2,3], [4,5,6]]})
+    obj = instantiate_quam_class(TestQuamSubList, {"sublist": [[1, 2, 3], [4, 5, 6]]})
 
-    assert obj.sublist == [[1,2,3], [4,5,6]]
+    assert obj.sublist == [[1, 2, 3], [4, 5, 6]]
+
+
+def test_instantiate_attr_literal():
+    attr = instantiate_attr(
+        attr_val="a",
+        expected_type=Literal["a", "b", "c"],
+    )
+    assert attr == "a"
+
+
+def test_instance_attr_literal_fail():
+    with pytest.raises(TypeError):
+        instantiate_attr(
+            attr_val="d",
+            expected_type=Literal["a", "b", "c"],
+        )
+
+    with pytest.raises(TypeError):
+        instantiate_attr(
+            attr_val=1,
+            expected_type=Literal["a", "b", "c"],
+        )