+++ tests/quam_base/test_quam_dict.py
@@ -201,3 +201,31 @@ def test_dict_unreferenced_value():
     assert d.val1 == 42
     assert d.val2 == 42
     assert d.get_unreferenced_value("val1") == "#./val2"
+
+
+def test_quam_dict_int_keys():
+    quam_dict = QuamDict({1: 2})
+    assert quam_dict.data == {1: 2}
+    assert quam_dict[1] == 2
+    quam_dict.pop(1)
+    assert quam_dict.data == {}
+    with pytest.raises(KeyError):
+        quam_dict[1]
+
+
+def test_quam_dict_get_attr_int():
+    quam_dict = QuamDict({1: 2})
+    assert quam_dict.get_attr_name(2) == 1
+
+
+def test_quam_dict_print_summary():
+    quam_dict = QuamDict({"a": "b", 1: 2})
+
+    from contextlib import redirect_stdout
+    import io
+
+    f = io.StringIO()
+    with redirect_stdout(f):
+        quam_dict.print_summary()
+    s = f.getvalue()
+    assert s == 'QuamDict (parent unknown):\n  a: "b"\n  1: 2\n'